Update node client script to create a ledger channel
